backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic-settings==2.6.1
SQLAlchemy[asyncio]==2.0.36
psycopg[binary]==3.2.3
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.1.1

